Reimplement ReadDir() on root FUSE node
